requirements.txt
gradio
gradio[oauth]
requests
google-genai
google-generativeai
googlesearch-python
pycountry
python-dotenv